chore(deps): upgrade wgpu 28→29, naga 28→29, jni 0.21→0.22
wgpu 29: Surface::get_current_texture() now returns CurrentSurfaceTexture
enum instead of Result. Updated both native (frame.rs) and wasm call sites
to match on the new variants. Also adapted pipeline.rs for Option-wrapped
depth stencil fields and bind_group_layouts, and init.rs for u64 limit type.

naga 29: No breaking public API changes for the GLSL/WGSL pipeline used here.

jni 0.22: Drop-in compatible; deprecated methods still functional.

+++ goud_engine/src/libs/graphics/backend/wgpu_backend/init.rs
@@ -90,7 +90,7 @@ impl WgpuBackend {
                 max_texture_units: limits.max_sampled_textures_per_shader_stage.min(16),
                 max_texture_size: limits.max_texture_dimension_2d,
                 max_vertex_attributes: limits.max_vertex_attributes,
-                max_uniform_buffer_size: limits.max_uniform_buffer_binding_size,
+                max_uniform_buffer_size: limits.max_uniform_buffer_binding_size as u32,
                 supports_instancing: true,
                 supports_compute_shaders: true,
                 supports_geometry_shaders: false,
